Revert "Create exam enhancement"
