--- .pre-commit-config.yaml
@@ -1,5 +1,0 @@
-repos:
-  - repo: https://github.com/qoomon/git-conventional-commits
-    rev: v2.6.3
-    hooks:
-      - id: conventional-commits
