+++ package.json
@@ -18,45 +18,45 @@
     "seed": "node prisma/seed.js"
   },
   "dependencies": {
-    "@hookform/resolvers": "^3.3.2",
+    "@hookform/resolvers": "^3.3.1",
     "@next-auth/prisma-adapter": "^1.0.7",
     "@pinecone-database/pinecone": "^0.1.6",
     "@prisma/client": "^4.16.2",
-    "@radix-ui/react-alert-dialog": "^1.0.5",
+    "@radix-ui/react-alert-dialog": "^1.0.4",
     "@radix-ui/react-aspect-ratio": "^1.0.3",
-    "@radix-ui/react-avatar": "^1.0.4",
+    "@radix-ui/react-avatar": "^1.0.3",
     "@radix-ui/react-checkbox": "^1.0.4",
-    "@radix-ui/react-context-menu": "^2.1.5",
-    "@radix-ui/react-dialog": "^1.0.5",
-    "@radix-ui/react-dropdown-menu": "^2.0.6",
+    "@radix-ui/react-context-menu": "^2.1.4",
+    "@radix-ui/react-dialog": "^1.0.4",
+    "@radix-ui/react-dropdown-menu": "^2.0.5",
     "@radix-ui/react-icons": "^1.3.0",
     "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-menubar": "^1.0.4",
-    "@radix-ui/react-popover": "^1.0.7",
+    "@radix-ui/react-menubar": "^1.0.3",
+    "@radix-ui/react-popover": "^1.0.6",
     "@radix-ui/react-progress": "^1.0.3",
     "@radix-ui/react-radio-group": "^1.1.3",
-    "@radix-ui/react-scroll-area": "^1.0.5",
+    "@radix-ui/react-scroll-area": "^1.0.4",
     "@radix-ui/react-select": "^1.2.2",
     "@radix-ui/react-separator": "^1.0.3",
     "@radix-ui/react-slider": "^1.1.2",
     "@radix-ui/react-slot": "^1.0.2",
     "@radix-ui/react-tabs": "^1.0.4",
-    "@radix-ui/react-toast": "^1.1.5",
+    "@radix-ui/react-toast": "^1.1.4",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@react-email/components": "^0.0.7",
     "@react-email/render": "^0.0.7",
     "@react-email/tailwind": "0.0.8",
-    "@sentry/nextjs": "^7.90.0",
+    "@sentry/nextjs": "^7.72.0",
     "@stripe/stripe-js": "^1.54.2",
     "@t3-oss/env-nextjs": "^0.2.2",
-    "@tanstack/react-query": "^4.36.1",
-    "@tanstack/react-table": "^8.11.0",
-    "@trpc/client": "^10.44.1",
-    "@trpc/next": "^10.44.1",
-    "@trpc/react-query": "^10.44.1",
-    "@trpc/server": "^10.44.1",
-    "@uploadthing/react": "^5.7.0",
-    "async": "^3.2.5",
+    "@tanstack/react-query": "^4.33.0",
+    "@tanstack/react-table": "^8.9.7",
+    "@trpc/client": "^10.38.1",
+    "@trpc/next": "^10.38.1",
+    "@trpc/react-query": "^10.38.1",
+    "@trpc/server": "^10.38.1",
+    "@uploadthing/react": "^5.5.0",
+    "async": "^3.2.4",
     "chalk": "^5.3.0",
     "class-variance-authority": "^0.6.1",
     "cli-progress": "^3.12.0",
@@ -71,25 +71,25 @@
     "lodash.chunk": "^4.2.0",
     "lucide-react": "^0.214.0",
     "micro": "^10.0.1",
-    "mixpanel-browser": "^2.48.1",
-    "next": "^13.5.6",
-    "next-auth": "^4.24.5",
+    "mixpanel-browser": "^2.47.0",
+    "next": "^13.4.19",
+    "next-auth": "^4.23.1",
     "next-i18next": "^13.3.0",
     "next-themes": "^0.2.1",
-    "nodemailer": "^6.9.7",
+    "nodemailer": "^6.9.5",
     "pdf-parse": "^1.1.1",
     "puppeteer": "^20.9.0",
     "react": "18.2.0",
-    "react-day-picker": "^8.9.1",
+    "react-day-picker": "^8.8.1",
     "react-dom": "18.2.0",
     "react-dom-confetti": "^0.2.0",
     "react-dropzone": "^14.2.3",
-    "react-email": "^1.10.0",
-    "react-error-boundary": "^4.0.12",
-    "react-hook-form": "^7.49.2",
+    "react-email": "^1.9.4",
+    "react-error-boundary": "^4.0.11",
+    "react-hook-form": "^7.46.1",
     "react-i18next": "^12.3.1",
     "react-markdown": "^8.0.7",
-    "recharts": "^2.10.3",
+    "recharts": "^2.8.0",
     "rehype-katex": "^6.0.3",
     "rehype-prism-plus": "^1.6.3",
     "rehype-raw": "^7.0.0",
@@ -102,35 +102,35 @@
     "superjson": "1.12.2",
     "tailwind-merge": "^1.14.0",
     "tailwindcss-animate": "^1.0.7",
-    "uploadthing": "^5.7.4",
-    "zod": "^3.22.4",
-    "zustand": "^4.4.7"
+    "uploadthing": "^5.5.3",
+    "zod": "^3.22.2",
+    "zustand": "^4.4.1"
   },
   "devDependencies": {
-    "@types/async": "^3.2.24",
+    "@types/async": "^3.2.20",
     "@types/chalk": "^2.2.0",
-    "@types/cli-progress": "^3.11.5",
-    "@types/eslint": "^8.56.0",
-    "@types/lodash.chunk": "^4.2.9",
-    "@types/lodash.partition": "^4.6.9",
-    "@types/mixpanel-browser": "^2.47.5",
-    "@types/node": "^18.19.3",
-    "@types/nodemailer": "^6.4.14",
+    "@types/cli-progress": "^3.11.2",
+    "@types/eslint": "^8.44.2",
+    "@types/lodash.chunk": "^4.2.7",
+    "@types/lodash.partition": "^4.6.7",
+    "@types/mixpanel-browser": "^2.47.1",
+    "@types/node": "^18.17.14",
+    "@types/nodemailer": "^6.4.9",
     "@types/prettier": "^2.7.3",
-    "@types/react": "^18.2.45",
-    "@types/react-dom": "^18.2.18",
-    "@types/sentiment": "^5.0.4",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "@types/sentiment": "^5.0.1",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
-    "autoprefixer": "^10.4.16",
-    "eslint": "^8.56.0",
-    "eslint-config-next": "^13.5.6",
-    "postcss": "^8.4.32",
+    "autoprefixer": "^10.4.15",
+    "eslint": "^8.48.0",
+    "eslint-config-next": "^13.4.19",
+    "postcss": "^8.4.29",
     "prettier": "^2.8.8",
     "prettier-plugin-tailwindcss": "^0.2.8",
     "prisma": "^4.16.2",
-    "tailwindcss": "^3.4.0",
-    "typescript": "^5.3.3"
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2"
   },
   "ct3aMetadata": {
     "initVersion": "7.13.0"
