+++ src/lib/ai/quiz.ts
@@ -53,11 +53,15 @@ const sentimentParser = StructuredOutputParser.fromZodSchema(
   })
 );
 
-export type SentimentParserResponseType = z.infer<
-  typeof sentimentParser.schema
->;
-export type QuizResponseType = z.infer<typeof quizParser.schema>;
-export type GradeQuizResponseType = z.infer<typeof gradeQuizParser.schema>;
+type SentimentParserResponseType = z.infer<typeof sentimentParser.schema>;
+type QuizResponseType = z.infer<typeof quizParser.schema>;
+type GradeQuizResponseType = z.infer<typeof gradeQuizParser.schema>;
+
+type ContextDocument = {
+  title: string;
+  url: string;
+  text: string;
+};
 
 // ------------------------------------ [Functions] ------------------------------------
 
@@ -71,13 +75,13 @@ export async function genQuiz(
         "Generate a SAT Practice Exam with {questions} questions."
       )
       .setRelevance(
-        "The exam should focus on the following subtopic: {subtopic}. As it pertains to your discipline.",
+        "The exam should focus on the subtopic: [{subtopic}] as it pertains to your discipline.",
         "Here are some user notes: {notes}"
       )
       .setConstraints(
         "Here are our supported question formats:",
         "Multiple Choice Questions (MCQ), Short Answer Questions (SA).",
-        "Difficulty of this exam should be {difficulty}"
+        "And should have a {difficulty} difficulty level"
       )
       .setPersonalization({ language: config.lang })
       .setRequirements();
@@ -166,7 +170,7 @@ export async function gradeQuiz(
 }
 
 export async function genReviewNotes(
-  results: Omit<GradeQuizResponseType[0], "correct">[],
+  results: GradeQuizResponseType,
   data: {
     subject: string;
     subtopic: string;
