+++ src/lib/mixpanel.tsx
@@ -17,7 +17,6 @@ export const mixpanelEventConfig = {
   Login: "Login" as const,
   Logout: "Logout" as const,
   Input: "Input" as const,
-  Error: "API Error" as const,
 };
 
 export type MixpanelPayload = {
