+++ src/server/api/routers/quiz.ts
@@ -2,15 +2,10 @@ import { z } from "zod";
 import sentiment from "sentiment";
 import { Role, Topics } from "@prisma/client";
 import { createTRPCRouter, protectedProcedure } from "@/server/api/trpc";
-import {
-  GradeQuizRequestType,
-  createQuizSchema,
-  gradeQuizSchema,
-} from "@/server/schemas";
+import { createQuizSchema, gradeQuizSchema } from "@/server/schemas";
 import { TRPCError } from "@trpc/server";
 
 import {
-  GradeQuizResponseType,
   genQuiz,
   genReviewNotes,
   gradeQuiz,
@@ -185,7 +180,12 @@ export const quizRouter = createTRPCRouter({
           });
         }
 
-        if (quiz.score > 0) return;
+        if (quiz.score > 0) {
+          throw new TRPCError({
+            code: "BAD_REQUEST",
+            message: "Quiz already graded",
+          });
+        }
 
         const result = await gradeQuiz(
           quiz.topic,
@@ -207,18 +207,29 @@ export const quizRouter = createTRPCRouter({
         const correctAnswers = result.filter((answer) => answer.correct);
         const score = Math.floor((correctAnswers.length / result.length) * 100);
 
-        await Promise.all([
-          ctx.prisma.quiz.update({
-            where: { id: input.quizId },
-            data: { score },
-          }),
-          sendEmail(ctx.session.user.email as string, "quizCompletion", {
-            name: ctx.session.user.name || "Anonymous",
-            score,
-            topic: quiz.topic,
+        const reviewNotes = await genReviewNotes(
+          result,
+          {
+            subject: quiz.topic,
             subtopic: quiz.subtopic,
-          }),
-        ]);
+            difficulty: quiz.difficulty,
+            score,
+          },
+          ctx.session.user.name || "Anonymous",
+          ctx.session.user.lang || "en"
+        );
+
+        await ctx.prisma.quiz.update({
+          where: { id: input.quizId },
+          data: { score, reviewNotes },
+        });
+
+        await sendEmail(ctx.session.user.email as string, "quizCompletion", {
+          name: ctx.session.user.name || "Anonymous",
+          score,
+          topic: quiz.topic,
+          subtopic: quiz.subtopic,
+        });
       } catch (e) {
         console.error(e);
         if (e instanceof TRPCError) throw e;
@@ -227,59 +238,6 @@ export const quizRouter = createTRPCRouter({
           message: "Failed to grade quiz",
         });
       }
-    }),
-  genReviewNotes: protectedProcedure
-    .input(z.object({ quizId: z.string() }))
-    .mutation(async ({ input, ctx }) => {
-      const quiz = await ctx.prisma.quiz.findUnique({
-        where: { id: input.quizId },
-        include: { questions: true },
-      });
-
-      if (!quiz) {
-        throw new TRPCError({
-          code: "NOT_FOUND",
-          message: "Quiz not found",
-        });
-      }
-
-      if (!quiz.score) {
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: "Quiz not graded",
-        });
-      }
-
-      if (!!quiz.reviewNotes) {
-        throw new TRPCError({
-          code: "BAD_REQUEST",
-          message: "Review notes already generated",
-        });
-      }
-
-      const results = quiz?.questions.map((q) => ({
-        question: q.label,
-        type: q.type,
-        studentAnswer: q.answer || "NOT ANSWERED",
-        answer: q.solution,
-      }));
-
-      const reviewNotes = await genReviewNotes(
-        results,
-        {
-          subject: quiz.topic,
-          subtopic: quiz.subtopic,
-          difficulty: quiz.difficulty,
-          score: quiz.score,
-        },
-        ctx.session.user.name || "Anonymous",
-        ctx.session.user.lang || "en"
-      );
-
-      await ctx.prisma.quiz.update({
-        where: { id: input.quizId },
-        data: { reviewNotes },
-      });
     }),
   getQuiz: protectedProcedure
     .input(
